+++ mlflow/__init__.py
@@ -105,7 +105,7 @@ from mlflow.config import (
 from mlflow.exceptions import MlflowException
 from mlflow.models import evaluate
 from mlflow.projects import run
-from mlflow.tracing.fluent import get_traces, start_span, trace
+from mlflow.tracing.python_api import start_span, trace
 from mlflow.tracking._model_registry.fluent import (
     register_model,
     search_model_versions,
